GXBABY USB driver

+++ drivers/usb/gadget/rndis.c
@@ -1063,7 +1063,7 @@ int rndis_rm_hdr(struct gether *port,
 
 		if (skb->len < sizeof *hdr) {
 			pr_err("invalid rndis pkt: skblen:%u hdr_len:%u",
-					skb->len, sizeof *hdr);
+					skb->len, (unsigned int)sizeof(*hdr));
 			dev_kfree_skb_any(skb);
 			return -EINVAL;
 		}
